Separates signage from inventory
