+++ README.md
@@ -158,13 +158,13 @@ The `/events` page shows the same calendar in **week** view.
 
 ## Inventory
 
-Visit `/inventario` to manage devices and road signage. Items open in modal
-dialogs for editing and each record includes a **quantità** field. The backend
-exposes Italian endpoint paths such as `/dispositivi`,
-`/inventario/signage-temp`, `/inventario/signage-vertical` and
-`/inventario/signage-horizontal`. Horizontal signage offers a **PDF anno** button
-that calls `/inventario/signage-horizontal/pdf?year=YYYY` to download the annual
-plan.
+Visit `/inventario` to manage only the available devices. Temporary and vertical
+signage have moved to the `/segnaletica` page. Items open in modal dialogs for
+editing and each record includes a **quantità** field. The backend exposes
+Italian endpoint paths such as `/dispositivi`, `/segnaletica-temporanea`,
+`/segnaletica-verticale` and `/inventario/signage-horizontal`. Horizontal
+signage offers a **PDF anno** button that calls
+`/inventario/signage-horizontal/pdf?year=YYYY` to download the annual plan.
 
 ## Segnalazioni
