+++ src/pages/ListPages.css
@@ -248,12 +248,12 @@
   display: flex;
   flex-wrap: wrap;
   gap: 1rem;
+  justify-content: center;
 }
 .inventory-wrapper > .temp-section,
 .inventory-wrapper > .vertical-section {
   flex: 1;
 }
 .inventory-wrapper > .devices-section {
-  order: 2;
-  flex-basis: 100%;
+  flex-basis: auto;
 }
